Added Dataset Investigator
The Dataset Investigator allows users to look through the batch displaying the images and the correct bounding boxes of the dataset the model will be trained with. Next up I'll try to restore the `global_step` variable when a checkpoint is restored, currently it always resets to 0 eventough a checkpoint is restored.