Added validation method
I moved a method for validating the trained network on a test image to the other files to make `main.py` less messy and the code easier to read. I changed from SGD to AdaGrad for testing purpose, I might also change the loss function since L2 loss is said to be inefficient especially for classification.